add leader broadcast_append_entries function
add this function that sends messages to each peer containing
the logs to be appended, keeping log consistency across nodes

add tests to this function

+++ src/raft/state_machine.rs
@@ -7,7 +7,7 @@ use log::info;
 // todo: update the command data structure
 #[derive(Debug)]
 pub struct Instruction {
-    pub index: u64,
+    pub index: usize,
     pub term: u64,
     pub command: String,
 }
@@ -15,7 +15,7 @@ pub struct Instruction {
 pub struct StateMachine {
     state_rx: UnboundedReceiverStream<Instruction>,
     node_tx: UnboundedSender<Message>,
-    applied_index: u64,
+    applied_index: usize,
 }
 
 impl StateMachine {
